chore: refresh project structure
