+++ Cargo.toml
@@ -1,17 +1,19 @@
 [package]
-    name = "rust_utils"
-    version = "0.1.0"
-    edition = "2021"
-    authors = ["Markus Swoveland"]
+name = "workflowpathoscope"
+version = "0.0.0"
+edition = "2021"
+authors = ["Markus Swoveland"]
+
+[lib]
+name = "workflow_pathoscope"
+# "cdylib" is necessary to produce a shared library for Python to import from.
+crate-type = ["cdylib"]
 
 [dependencies]
 yaml-rust = "0.4.5"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies.pyo3]
-version = "0.16.3"
-
-[features]
-extension-module = ["pyo3/extension-module"]
-default = ["extension-module"]
+version = "^0.20.0"
+# "abi3-py38" tells pyo3 (and maturin) to build using the stable ABI with minimum
+# Python version 3.12
+features = ["abi3-py312"]
