+++ python/workflow_pathoscope/__init__.py
@@ -1,0 +1,3 @@
+from workflow_pathoscope import workflow_pathoscope as rust
+
+
