+++ workflow.py
@@ -8,7 +8,6 @@ from typing import Any, TextIO
 
 import aiofiles
 import aiofiles.os
-import rust_utils
 from virtool_workflow import hooks, step
 from virtool_workflow.data.analyses import WFAnalysis
 from virtool_workflow.data.indexes import WFIndex
@@ -16,7 +15,7 @@ from virtool_workflow.data.samples import WFSample
 from virtool_workflow.data.subtractions import WFSubtraction
 from virtool_workflow.runtime.run_subprocess import RunSubprocess
 
-from utils import SamLine, calculate_coverage, run_pathoscope, write_report
+from workflow_pathoscope import SamLine, calculate_coverage, run_pathoscope, write_report
 
 BAD_FIRST_SAM_CHARACTERS = {"\n", "@", "#"}
